1.3-r3 + Created SideWidget class inherited by left, right and loader widgets + removed icons from old version
